+++ src/nos/helmholtz/__init__.py
@@ -1,4 +1,7 @@
 from .dataset import HelmholtzDataset
 from .helmholtz import Helmholtz
 
-__all__ = ["Helmholtz", "HelmholtzDataset"]
+__all__ = [
+    "Helmholtz",
+    "HelmholtzDataset",
+]
