+++ src/nos/helmholtz/domain_properties/absorber_description.py
