+++ src/nos/helmholtz/domain_properties/crystal_description.py
