+++ src/nos/helmholtz/domain_properties/domain_description.py
