+++ src/nos/helmholtz/solver/wave_number_function.py
